query builder: cleanup query processor and builder logic
