+++ packages/legend-application-query/src/stores/fetch-structure/graph-fetch/QueryBuilderGraphFetchTreeStateBuilder.ts
@@ -1,0 +1,127 @@
+/**
+ * Copyright (c) 2020-present, Goldman Sachs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {
+  GraphFetchTreeInstanceValue,
+  matchFunctionName,
+  RootGraphFetchTree,
+  SimpleFunctionExpression,
+} from '@finos/legend-graph';
+import { assertTrue, guaranteeType } from '@finos/legend-shared';
+import { QUERY_BUILDER_SUPPORTED_FUNCTIONS } from '../../../QueryBuilder_Const.js';
+import type { QueryBuilderState } from '../../QueryBuilderState.js';
+import { QueryBuilderValueSpecificationProcessor } from '../../QueryBuilderStateBuilder.js';
+import { FETCH_STRUCTURE_IMPLEMENTATION } from '../QueryBuilderFetchStructureImplementationState.js';
+import { QueryBuilderGraphFetchTreeState } from './QueryBuilderGraphFetchTreeState.js';
+import { buildGraphFetchTreeData } from './QueryBuilderGraphFetchTreeUtil.js';
+
+export const processGraphFetchExpression = (
+  expression: SimpleFunctionExpression,
+  queryBuilderState: QueryBuilderState,
+): void => {
+  const functionName = expression.functionName;
+
+  assertTrue(
+    expression.parametersValues.length === 2,
+    `Can't process ${functionName}() expression: ${functionName}() expects 1 argument`,
+  );
+
+  const precedingExpression = guaranteeType(
+    expression.parametersValues[0],
+    SimpleFunctionExpression,
+    `Can't process ${functionName}() expression: only support ${functionName}() immediately following an expression`,
+  );
+  precedingExpression.accept_ValueSpecificationVisitor(
+    new QueryBuilderValueSpecificationProcessor(queryBuilderState, expression),
+  );
+
+  // check caller
+  assertTrue(
+    [
+      QUERY_BUILDER_SUPPORTED_FUNCTIONS.FILTER,
+      QUERY_BUILDER_SUPPORTED_FUNCTIONS.GET_ALL,
+    ].some((fn) => matchFunctionName(precedingExpression.functionName, fn)),
+    `Can't process ${functionName}(): only support ${functionName}() immediately following either getAll() or filter()`,
+  );
+
+  // build state
+  if (
+    queryBuilderState.fetchStructureState.implementation instanceof
+    QueryBuilderGraphFetchTreeState
+  ) {
+    const graphFetchTreeState =
+      queryBuilderState.fetchStructureState.implementation;
+    graphFetchTreeState.setChecked(
+      matchFunctionName(
+        expression.functionName,
+        QUERY_BUILDER_SUPPORTED_FUNCTIONS.GRAPH_FETCH_CHECKED,
+      ),
+    );
+  }
+};
+
+export const processGraphFetchSerializeExpression = (
+  expression: SimpleFunctionExpression,
+  queryBuilderState: QueryBuilderState,
+): void => {
+  // update fetch-structure
+  queryBuilderState.fetchStructureState.changeImplementation(
+    FETCH_STRUCTURE_IMPLEMENTATION.GRAPH_FETCH,
+  );
+
+  assertTrue(
+    expression.parametersValues.length === 2,
+    `Can't process serialize() expression: serialize() expects 1 argument`,
+  );
+
+  const precedingExpression = guaranteeType(
+    expression.parametersValues[0],
+    SimpleFunctionExpression,
+    `Can't process serialize() expression: only support serialize() immediately following an expression`,
+  );
+  precedingExpression.accept_ValueSpecificationVisitor(
+    new QueryBuilderValueSpecificationProcessor(queryBuilderState, expression),
+  );
+
+  // check caller
+  assertTrue(
+    [
+      QUERY_BUILDER_SUPPORTED_FUNCTIONS.GRAPH_FETCH,
+      QUERY_BUILDER_SUPPORTED_FUNCTIONS.GRAPH_FETCH_CHECKED,
+    ].some((fn) => matchFunctionName(precedingExpression.functionName, fn)),
+    `Can't process serialize() expression: only support serialize() in graph-fetch expression`,
+  );
+
+  // build state
+  if (
+    queryBuilderState.fetchStructureState.implementation instanceof
+    QueryBuilderGraphFetchTreeState
+  ) {
+    const graphFetchTreeState =
+      queryBuilderState.fetchStructureState.implementation;
+    const serializeFunc = guaranteeType(
+      expression.parametersValues[1],
+      GraphFetchTreeInstanceValue,
+      `Can't process serialize() expression: serialize() graph-fetch is missing`,
+    );
+    const value = guaranteeType(
+      serializeFunc.values[0],
+      RootGraphFetchTree,
+      `Can't process serialize() expression: serialize() graph-fetch tree root is missing`,
+    );
+    graphFetchTreeState.setGraphFetchTree(buildGraphFetchTreeData(value));
+  }
+};
